{"schema":"decodex/commit/1","summary":"Split quantitative product row validation","authority":"manual"}
